Add files via upload
These are the console view and the console controller as well as a demo console that still needs work. The code for the console view and console controller is from https://www.packtpub.com/books/content/making-game-console-unity-part-2. With the console controller, I tried to recreate the code I found, but it didn't work. The code is commented out but when merging everything, it didn't quite work right.